refactor(db): extract app_settings and repository_mcp_server into db/settings.rs
